# yoru-platform/openapi/yoru-api-outline.yaml
openapi: 3.1.0
info:
  title: Yoru API
  version: 0.2.1-sprint1
  summary: Programming contract outline for the Yoru platform
  description: >
    API outline with implemented Sprint 1 identity contracts and representative
    contracts for later business modules.
servers:
  - url: /api/v1
    description: Environment-relative API base path
...
security:
  - sessionCookie: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a customer account
      operationId: registerCustomer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      tags: [Auth]
      summary: Create a browser session
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session cookies set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the current refresh token
      operationId: refreshSession
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Rotated session cookies set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session revoked
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user and capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Revoke every session owned by the current user
      operationId: logoutAll
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: All sessions revoked
  /auth/active-partner:
    patch:
      tags: [Auth]
      summary: Select an active partner from the current user's memberships
      operationId: selectActivePartner
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectPartnerRequest"
      responses:
        "200":
          description: Active tenant changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /auth/sessions/{session_id}:
    delete:
      tags: [Auth]
      summary: Revoke a session owned by the current user
      operationId: revokeOwnSession
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session revoked
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yoru_access
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 20
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Permission, tenant, state, CSRF, or MFA policy denied the request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found or hidden by access policy
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Resource state, concurrency, or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 12
          maxLength: 128
        full_name:
          type: string
          minLength: 2
          maxLength: 150
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          maxLength: 128
    User:
      type: object
      required: [id, email, full_name, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        status:
          type: string
          enum: [active, suspended, blocked]
    Membership:
      type: object
      required: [partner_id, partner_status, membership_status, role, permissions]
      properties:
        partner_id:
          type: string
          format: uuid
        partner_status:
          type: string
        membership_status:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
    SessionResponse:
      type: object
      required: [user, active_partner_id, platform_roles, permissions, memberships]
      properties:
        user:
          $ref: "#/components/schemas/User"
        active_partner_id:
          type: [string, "null"]
          format: uuid
        platform_roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
    SelectPartnerRequest:
      type: object
      additionalProperties: false
      required: [partner_id]
      properties:
        partner_id:
          type: [string, "null"]
          format: uuid
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              field:
                type: [string, "null"]
              code:
                type: string
              message:
                type: string
...